Remove EDH address api methods & tests
