+++ source/api/address/__tests__/address--test.js
@@ -1,93 +1,16 @@
-import moxios from 'moxios'
 import { searchAddress, getAddressDetails } from '..'
-import { instance, servicesAPI, updateClient } from '../../../utils/client'
+import { servicesAPI } from '../../../utils/client'
 
-describe('Address search', () => {
-  describe('Search Address EDH', () => {
-    beforeEach(() => {
-      moxios.install(instance)
-    })
-
-    afterEach(() => {
-      moxios.uninstall(instance)
-    })
-
-    it('uses the correct url', done => {
-      searchAddress('Foo street')
-
-      moxios.wait(() => {
-        const request = moxios.requests.mostRecent()
-        expect(request.url).to.contain(
-          'https://everydayhero.com/api/v2/addresses'
-        )
-        expect(request.url).to.contain('q=Foo+street')
-        expect(request.url).to.contain('country_code=au')
-        done()
-      })
-    })
-
-    it('uses the correct url for a different region', done => {
-      searchAddress('Foo street', 'us')
-
-      moxios.wait(() => {
-        const request = moxios.requests.mostRecent()
-        expect(request.url).to.contain(
-          'https://everydayhero.com/api/v2/addresses'
-        )
-        expect(request.url).to.contain('q=Foo+street')
-        expect(request.url).to.contain('country_code=us')
-        done()
-      })
-    })
+describe('Addresses', () => {
+  beforeEach(() => {
+    moxios.install(servicesAPI)
   })
 
-  describe('Get Address Details EDH', () => {
-    beforeEach(() => {
-      moxios.install(instance)
-    })
-
-    afterEach(() => {
-      moxios.uninstall(instance)
-    })
-
-    it('uses the correct url', done => {
-      getAddressDetails(123)
-
-      moxios.wait(() => {
-        const request = moxios.requests.mostRecent()
-        expect(request.url).to.contain(
-          'https://everydayhero.com/api/v2/addresses/au/123'
-        )
-        done()
-      })
-    })
-
-    it('uses the correct url for a different region', done => {
-      getAddressDetails(123, 'us')
-
-      moxios.wait(() => {
-        const request = moxios.requests.mostRecent()
-        expect(request.url).to.contain(
-          'https://everydayhero.com/api/v2/addresses/us/123'
-        )
-        done()
-      })
-    })
+  afterEach(() => {
+    moxios.uninstall(servicesAPI)
   })
 
-  describe('Search Address JG', () => {
-    beforeEach(() => {
-      updateClient({ baseURL: 'https://api.justgiving.com' })
-      moxios.install(instance)
-      moxios.install(servicesAPI)
-    })
-
-    afterEach(() => {
-      updateClient({ baseURL: 'https://everydayhero.com' })
-      moxios.uninstall(instance)
-      moxios.uninstall(servicesAPI)
-    })
-
+  describe('Address search', () => {
     it('uses the correct url', done => {
       searchAddress('SW1A1AA')
 
@@ -102,18 +25,6 @@ describe('Address search', () => {
   })
 
   describe('Get Address Details JG', () => {
-    beforeEach(() => {
-      updateClient({ baseURL: 'https://api.justgiving.com' })
-      moxios.install(instance)
-      moxios.install(servicesAPI)
-    })
-
-    afterEach(() => {
-      updateClient({ baseURL: 'https://everydayhero.com' })
-      moxios.uninstall(instance)
-      moxios.uninstall(servicesAPI)
-    })
-
     it('uses the correct url', done => {
       getAddressDetails('1234')
